rtc uses uint64_t for addressing

+++ pyboy/core/cartridge/rtc.pxd
@@ -33,6 +33,6 @@ cdef class RTC:
     @cython.locals(days=uint64_t)
     cdef void latch_rtc(self) noexcept nogil
     cdef void writecommand(self, uint8_t) noexcept nogil
-    cdef uint8_t getregister(self, uint8_t) noexcept nogil
+    cdef uint8_t getregister(self, uint64_t) noexcept nogil
     @cython.locals(t=cython.double, days=uint64_t)
     cdef void setregister(self, uint8_t, uint8_t) noexcept nogil
